# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Mnemos API",
  "description": "Governed long-term memory and continuous learning for AI agents.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/conflicts": {
   "get": {
    "tags": [
     "conflicts"
    ],
    "summary": "List Conflicts",
    "operationId": "list_conflicts_v1_conflicts_get",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Workspace Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Conflicts V1 Conflicts Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/conflicts/{conflict_id}": {
   "get": {
    "tags": [
     "conflicts"
    ],
    "summary": "Get Conflict",
    "operationId": "get_conflict_v1_conflicts__conflict_id__get",
    "parameters": [
     {
      "name": "conflict_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Conflict Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConflictDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/conflicts/{conflict_id}/resolve": {
   "post": {
    "tags": [
     "conflicts"
    ],
    "summary": "Resolve",
    "operationId": "resolve_v1_conflicts__conflict_id__resolve_post",
    "parameters": [
     {
      "name": "conflict_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Conflict Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ResolveIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConflictDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/dreams": {
   "post": {
    "tags": [
     "dreams"
    ],
    "summary": "Request Dream",
    "operationId": "request_dream_v1_dreams_post",
    "parameters": [
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DreamIn"
       }
      }
     }
    },
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DreamOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "dreams"
    ],
    "summary": "List Dreams",
    "operationId": "list_dreams_v1_dreams_get",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Workspace Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "cursor",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "uuid"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Dreams V1 Dreams Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/dreams/{dream_id}": {
   "get": {
    "tags": [
     "dreams"
    ],
    "summary": "Get Dream",
    "operationId": "get_dream_v1_dreams__dream_id__get",
    "parameters": [
     {
      "name": "dream_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid",
       "title": "Dream Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     },
     {
      "name": "x-api-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Api-Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DreamOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ConflictDetail": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "workspace_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Workspace Id"
     },
     "candidate_memory_id": {
      "type": "string",
      "format": "uuid",
      "title": "Candidate Memory Id"
     },
     "existing_memory_id": {
      "type": "string",
      "format": "uuid",
      "title": "Existing Memory Id"
     },
     "conflict_type": {
      "type": "string",
      "title": "Conflict Type"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "analysis_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Analysis Json"
     },
     "resolution_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Resolution Json"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "resolved_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Resolved At"
     },
     "candidate": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/MemoryOut"
       },
       {
        "type": "null"
       }
      ]
     },
     "existing": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/MemoryOut"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "workspace_id",
     "candidate_memory_id",
     "existing_memory_id",
     "conflict_type",
     "status",
     "analysis_json",
     "resolution_json",
     "created_at",
     "resolved_at"
    ],
    "title": "ConflictDetail"
   },
   "ConflictResolution": {
    "type": "string",
    "enum": [
     "keep_existing",
     "accept_candidate",
     "keep_both",
     "archive_both"
    ],
    "title": "ConflictResolution"
   },
...
   "DreamIn": {
    "properties": {
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "mode": {
      "$ref": "#/components/schemas/DreamMode"
     },
     "dedupe_window": {
      "type": "boolean",
      "title": "Dedupe Window",
      "default": false
     }
    },
    "type": "object",
    "required": [
     "workspace_id",
     "mode"
    ],
    "title": "DreamIn"
   },
   "DreamMode": {
    "type": "string",
    "enum": [
     "reflection",
     "deduplication",
     "pattern",
     "contradiction",
     "generalization",
     "compression"
    ],
    "title": "DreamMode"
   },
   "DreamOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "workspace_id": {
      "type": "string",
      "format": "uuid",
      "title": "Workspace Id"
     },
     "mode": {
      "type": "string",
      "title": "Mode"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "trigger_type": {
      "type": "string",
      "title": "Trigger Type"
     },
     "window_hash": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Window Hash"
     },
     "input_window_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Input Window Json"
     },
     "result_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Result Json"
     },
     "checkpoint_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Checkpoint Json"
     },
     "error": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Error"
     },
     "started_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     },
     "completed_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Completed At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "workspace_id",
     "mode",
     "status",
     "trigger_type",
     "window_hash",
     "input_window_json",
     "result_json",
     "checkpoint_json",
     "error",
     "started_at",
     "completed_at",
     "created_at"
    ],
    "title": "DreamOut"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MemoryOut": {
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid",
      "title": "Id"
     },
     "organization_id": {
      "type": "string",
      "format": "uuid",
      "title": "Organization Id"
     },
     "workspace_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Workspace Id"
     },
     "project_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Project Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string",
        "format": "uuid"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "layer": {
      "type": "integer",
      "title": "Layer"
     },
     "type": {
      "type": "string",
      "title": "Type"
     },
     "scope_type": {
      "type": "string",
      "title": "Scope Type"
     },
     "scope_id": {
      "type": "string",
      "format": "uuid",
      "title": "Scope Id"
     },
     "title": {
      "type": "string",
      "title": "Title"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "review_state": {
      "type": "string",
      "title": "Review State"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "trust_score": {
      "type": "number",
      "title": "Trust Score"
     },
     "importance": {
      "type": "number",
      "title": "Importance"
     },
     "utility_score": {
      "type": "number",
      "title": "Utility Score"
     },
     "valid_from": {
      "type": "string",
      "format": "date-time",
      "title": "Valid From"
     },
     "valid_until": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Valid Until"
     },
     "version": {
      "type": "integer",
      "title": "Version"
     },
     "metadata_json": {
      "additionalProperties": true,
      "type": "object",
      "title": "Metadata Json"
     },
     "retrieval_count": {
      "type": "integer",
      "title": "Retrieval Count"
     },
     "last_retrieved_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Retrieved At"
     },
     "created_by_type": {
      "type": "string",
      "title": "Created By Type"
     },
     "created_by_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Created By Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "organization_id",
     "workspace_id",
     "project_id",
     "agent_id",
     "layer",
     "type",
     "scope_type",
     "scope_id",
     "title",
     "content",
     "status",
     "review_state",
     "confidence",
     "trust_score",
     "importance",
     "utility_score",
     "valid_from",
     "valid_until",
     "version",
     "metadata_json",
     "retrieval_count",
     "last_retrieved_at",
     "created_by_type",
     "created_by_id",
     "created_at",
     "updated_at"
    ],
    "title": "MemoryOut"
   },
...
   "ResolveIn": {
    "properties": {
     "resolution": {
      "$ref": "#/components/schemas/ConflictResolution"
     },
     "note": {
      "type": "string",
      "maxLength": 2000,
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "resolution"
    ],
    "title": "ResolveIn"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}